Reduce scope of resource ordering for User, Group and File
Without this change, this module causes *every* user and group resource to
come before any file resource.  This is a problem because it is likely to
cause circular dependencies when this module is used in a larger context.
This change fixes that by limiting the dependencies to just the resources
within the module that require it.